Componentize transactions
